@@ -22,14 +22,15 @@
     "type": "git",
     "url": "https://github.com/ksv741/eslint-config-ksv741"
   },
+  "type": "module",
   "version": "0.1.0",
   "devDependencies": {
     "@commitlint/cli": "19.5.0",
     "@commitlint/config-conventional": "19.5.0",
-    "eslint": "9.13.0",
+    "eslint": "9.31.0",
     "husky": "9.1.6",
     "standard-version": "9.5.0",
-    "typescript": "5.6.3"
+    "typescript": "5.8.3"
   },
   "main": "index.js",
   "files": [
@@ -41,22 +42,22 @@
     "node": ">=16.13.0"
   },
   "dependencies": {
-    "@eslint/eslintrc": "3.1.0",
-    "@stylistic/eslint-plugin": "2.10.0",
-    "eslint-import-resolver-typescript": "3.6.3",
-    "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-jest": "28.8.3",
-    "eslint-plugin-jest-dom": "5.4.0",
+    "@eslint/eslintrc": "3.3.1",
+    "@stylistic/eslint-plugin": "5.2.0",
+    "eslint-import-resolver-typescript": "4.4.4",
+    "eslint-plugin-import": "2.32.0",
+    "eslint-plugin-jest": "29.0.1",
+    "eslint-plugin-jest-dom": "5.5.0",
     "eslint-plugin-jest-formatting": "3.1.0",
     "eslint-plugin-jsx-a11y": "6.10.2",
-    "eslint-plugin-react": "7.37.2",
-    "eslint-plugin-react-hooks": "5.0.0",
-    "eslint-plugin-testing-library": "6.4.0",
-    "globals": "15.11.0",
-    "typescript-eslint": "8.12.2"
+    "eslint-plugin-react": "7.37.5",
+    "eslint-plugin-react-hooks": "5.2.0",
+    "eslint-plugin-testing-library": "7.6.0",
+    "globals": "16.3.0",
+    "typescript-eslint": "8.37.0"
   },
   "peerDependencies": {
-    "eslint": "^9.13.0",
+    "eslint": "^9.31.0",
     "typescript": ">=4.2.2"
   },
   "peerDependenciesMeta": {
