@@ -1,5 +1,9 @@
-module.exports = [
-  require('../rules/.eslint-js'),
-  require('../rules/.eslint-import'),
-  require('../rules/style/.eslint-style-js'),
+import jsBaseConfig from '../rules/.eslint-js.js'
+import importConfig from '../rules/.eslint-import.js'
+import styleJsConfig from '../rules/style/.eslint-style-js.js';
+
+export default [
+  jsBaseConfig,
+  importConfig,
+  styleJsConfig,
 ];
