@@ -1,5 +1,9 @@
-module.exports = [
-  ...require('./js-base'),
-  require('../rules/.eslint-ts'),
-  require('../rules/style/.eslint-style-ts'),
+import jsBaseConfig from './js-base.js';
+import eslintTsBaseConfig from '../rules/.eslint-ts.js';
+import styleTsConfig from '../rules/style/.eslint-style-ts.js';
+
+export default [
+  ...jsBaseConfig,
+  eslintTsBaseConfig,
+  styleTsConfig,
 ];
