@@ -1,6 +1,6 @@
-const stylistic = require('@stylistic/eslint-plugin');
+import stylistic from '@stylistic/eslint-plugin';
 
-module.exports = {
+export default {
   name: 'style-ts',
   plugins: {
     '@stylistic': stylistic,
